Object Schema Relations built & gradle error fixed

+++ app/src/main/java/com/example/splitapp/db/Bill.java
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Bill {
     @PrimaryKey(autoGenerate = true)
-    public int billID;
+    public long billID;
 
     @ColumnInfo(name = "name")
     public String name;
