+++ app/src/main/java/com/example/splitapp/db/Item.java
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Item {
     @PrimaryKey(autoGenerate = true)
-    public int itemID;
+    public long itemID;
     @ColumnInfo(name = "name")
     public String name;
     @ColumnInfo(name = "cost")
